refactor public content contracts
Move pure public content defaults, normalizers, validators, page-key helpers, and homepage summary helpers into @anipotts/content/public. Keep @anipotts/lib/cms as the D1-backed reader facade and compatibility export.

Update public-site consumers, content metadata labels, deploy target calculation, and sensitive path review coverage. Add a metadata-only D1 migration to refresh inert content operation source_ref values after merge.

Verified: pnpm --filter @anipotts/content typecheck; pnpm --filter @anipotts/content build; pnpm --filter @anipotts/lib typecheck; pnpm --filter @anipotts/lib build; pnpm --filter @anipotts/lib test -- --run src/cms/index.test.ts; pnpm turbo typecheck --filter=@anipotts/www...; pnpm turbo build --filter=@anipotts/www...; pnpm turbo typecheck --filter=@anipotts/admin...; pnpm turbo build --filter=@anipotts/admin...; pnpm test:content-platform; pnpm test:public-boundary; pnpm test:deploy-targets; pnpm test:security-review; pnpm test:public-routes; pnpm test:admin-routes; pnpm format:check; git diff --check; pnpm --silent proof:admin-content; pnpm --silent proof:admin-passkey.
