+++ apps/www/src/components/HomepageRichSummary.astro
@@ -5,7 +5,7 @@ import type {
   HomepageRichSummarySegment,
   HomepageRichSummarySentence,
   HomepageRichSummarySimpleSegment,
-} from "@anipotts/lib/cms";
+} from "@anipotts/content/public";
 
 interface Props {
   sentences: HomepageRichSummarySentence[];
