+++ apps/www/src/pages/making.astro
@@ -6,10 +6,10 @@ import { setDB } from "@anipotts/lib/db";
 import type { D1Database } from "@anipotts/lib/db";
 import {
   DEFAULT_MAKING_INDEX_CONTENT,
-  fetchPageContent,
   normalizeListingPageContent,
   type ListingPageContent,
-} from "@anipotts/lib/cms";
+} from "@anipotts/content/public";
+import { fetchPageContent } from "@anipotts/lib/cms";
 
 export const prerender = false;
