+++ packages/content/package.json
@@ -11,6 +11,22 @@
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js"
     },
+    "./public": {
+      "types": "./dist/public/index.d.ts",
+      "import": "./dist/public/index.js"
+    },
+    "./public/defaults": {
+      "types": "./dist/public/defaults.d.ts",
+      "import": "./dist/public/defaults.js"
+    },
+    "./public/editor": {
+      "types": "./dist/public/editor.d.ts",
+      "import": "./dist/public/editor.js"
+    },
+    "./public/homepage": {
+      "types": "./dist/public/homepage.d.ts",
+      "import": "./dist/public/homepage.js"
+    },
     "./admin": {
       "types": "./dist/admin/index.d.ts",
       "import": "./dist/admin/index.js"
@@ -20,6 +36,9 @@
     "build": "tsc",
     "dev": "tsc --watch",
     "typecheck": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@anipotts/types": "workspace:*"
   },
   "devDependencies": {
     "typescript": "^5"
