+++ packages/content/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./admin/index.js";
+export * from "./public/index.js";
