+++ packages/content/src/public/defaults.ts
@@ -1,0 +1,499 @@
+import type {
+  CmsProjectContent,
+  CmsWritingContent,
+  HomepageContent,
+  ListingPageContent,
+  NewsletterContent,
+  OrchestratingPageContent,
+} from "@anipotts/types";
+
+export const HOME_SECTION_ORDER: HomepageContent["section_order"] = [
+  "intro",
+  "about",
+  "past_work",
+  "latest_thoughts",
+];
+
+export const HOMEPAGE_FIELD_LIMITS = {
+  label: 80,
+  heading: 160,
+  subheading: 500,
+  paragraph: 1200,
+  linkLabel: 80,
+  linkHref: 300,
+  mentionKey: 80,
+  mentionLabel: 120,
+  mentionAsset: 220,
+  slug: 120,
+  proofTitle: 120,
+  proofDetail: 360,
+  limitMin: 1,
+  limitMax: 12,
+} as const;
+
+export const DEFAULT_HOMEPAGE_CONTENT: HomepageContent = {
+  sections: {
+    intro: {
+      visible: true,
+      label: "index",
+      heading: "hi, i'm ani",
+      subheading:
+        "previously worked on real-time agent i/o at structured ai (YC F25) and our bad habit, an atlantic records venture. every now and then i post about what i'm doing with claude code and codex.",
+      rich_summary: [
+        {
+          segments: [
+            {
+              kind: "text",
+              text: "previously worked on real-time agent i/o at ",
+            },
+            {
+              kind: "cluster",
+              segments: [
+                { kind: "mention", key: "structuredAi" },
+                {
+                  kind: "parens",
+                  segments: [{ kind: "mention", key: "yCombinatorF25" }],
+                },
+              ],
+            },
+            { kind: "text", text: " and " },
+            {
+              kind: "cluster",
+              segments: [
+                { kind: "mention", key: "badHabit", suffix: "," },
+                { kind: "text", text: "an " },
+                { kind: "mention", key: "atlanticRecords" },
+                { kind: "text", text: " venture." },
+              ],
+            },
+          ],
+        },
+        {
+          segments: [
+            {
+              kind: "text",
+              text: "every now and then i post about what i'm doing with claude code and codex, as featured on ",
+            },
+            { kind: "mention", key: "businessInsider", suffix: "." },
+          ],
+        },
+      ],
+    },
+    about: {
+      visible: true,
+      label: "about",
+      heading: "",
+      paragraphs: [
+        "I design multi-agent systems that orchestrate combinatorial verification tasks. Autonomous QA/QC pipelines for AEC, SEO auditing at scale, document extraction across regulatory domains. Previously full stack at a YC F25 startup (under NDA).",
+        "Everything I build ships with Claude Code. I publish the workflows, tooling, and real usage data as I go. Recently featured in Business Insider on how developers are restructuring their days around AI tools.",
+      ],
+    },
+    past_work: {
+      visible: true,
+      label: "making",
+      heading: "",
+      limit: 4,
+      links: [{ label: "view all", href: "/making" }],
+      view_all: "/making",
+      project_slugs: [
+        "quantercise",
+        "quantercise-extension",
+        "saeshify",
+        "nyu-purity-test",
+      ],
+    },
+    latest_thoughts: {
+      visible: true,
+      label: "writing",
+      heading: "",
+      limit: 3,
+      links: [{ label: "view all", href: "/writing" }],
+      view_all: "/writing",
+      writing_slugs: [
+        "saturdays-are-for-claude-code",
+        "i-built-a-monitor-for-my-claude-code-sessions",
+        "stop-ending-your-day-with-fix-the-bug",
+      ],
+    },
+  },
+  section_order: HOME_SECTION_ORDER,
+  mentions: {
+    structuredAi: {
+      label: "structured ai",
+      href: "https://getstructured.ai/",
+      logoSrc: "/images/brand/structured-ai-favicon.png",
+      logoAlt: "structured ai",
+      logoTone: "white",
+    },
+    yCombinatorF25: {
+      label: "F25",
+      logoSrc: "/images/brand/ycombinator-favicon.ico",
+      logoAlt: "y combinator",
+    },
+    badHabit: {
+      label: "our bad habit",
+      href: "https://ourbadhabit.com/",
+      logoSrc: "/images/brand/bad-habit-favicon.png",
+      logoAlt: "our bad habit",
+    },
+    atlanticRecords: {
+      label: "atlantic records",
+      href: "https://www.atlanticrecords.com/",
+      logoSrc: "/images/brand/atlantic-records-logo-cropped.png",
+      logoAlt: "atlantic records",
+      logoShape: "wide",
+    },
+    businessInsider: {
+      label: "business insider",
+      href: "/writing/saturdays-are-for-claude-code",
+      logoSrc: "/images/brand/business-insider-favicon.svg",
+      logoAlt: "business insider",
+    },
+  },
+  proof_cards: [
+    {
+      label: "structured ai",
+      href: "https://getstructured.ai/",
+      title: "drawing chat with page-level citations",
+      detail:
+        "architectural PDFs in, cited answers out. streamed claude/gemini and kept redis replay for live spectating and later debugging.",
+    },
+    {
+      label: "quantercise",
+      href: "https://quantercise.com",
+      title: "quant prep with real grading",
+      detail:
+        "next.js, typescript, postgres, drizzle, stripe, sandboxed python, and math-heavy grading paths.",
+    },
+    {
+      label: "paragon global investments",
+      href: "https://paragoninvestments.org",
+      title: "research portal for a quant fund",
+      detail:
+        "next.js and typescript on supabase. made fund research searchable and usable from mobile instead of buried in scattered docs.",
+    },
+    {
+      label: "public tooling",
+      href: "/projects/claude-code-tips",
+      title: "tools i actually use",
+      detail:
+        "claude-code-tips and imessage mcp are small public receipts from the same local-first workflow i run every day.",
+    },
+  ],
+};
+
+const CMS_LINK_LIMIT = 300;
+
+export const CMS_TEXT_LIMITS = {
+  slug: 120,
+  title: 160,
+  status: 40,
+  year: 40,
+  range: 80,
+  tag: 40,
+  summary: 600,
+  body: 12000,
+  linkLabel: 80,
+  linkUrl: CMS_LINK_LIMIT,
+  newsletterHeadline: 160,
+  newsletterDeck: 600,
+  newsletterFooter: 1200,
+  sender: 120,
+} as const;
+
+export const DEFAULT_NEWSLETTER_CONTENT: NewsletterContent = {
+  headline: "notes from the build loop",
+  deck: "notes on agent workflows and product builds, including the parts that broke while shipping.",
+  cta_label: "subscribe",
+  success_message: "subscribed. check your inbox.",
+  error_message: "could not subscribe. try again in a minute.",
+  footer_text: "you can unsubscribe at any time.",
+  buttondown_url: "https://news.anipotts.com",
+  archive_label: "archive",
+  archive_copy: "published notes live here as the archive fills in.",
+  archive_link_label: "read the archive",
+  archive_url: "/archive",
+  sender_name: "Ani Potts",
+  sender_email: "news@anipotts.com",
+  reply_to: "contact@anipotts.com",
+};
+
+export const DEFAULT_WRITING_INDEX_CONTENT: ListingPageContent = {
+  title: "writing",
+  description:
+    "things ani's written. claude code workflows, ai builds, the occasional music take.",
+  hero_title: "writing",
+  hero_summary: "stuff i've figured out and wanted to write down.",
+  search_placeholder: "search writing",
+};
+
+export const DEFAULT_MAKING_INDEX_CONTENT: ListingPageContent = {
+  title: "making",
+  description:
+    "projects, experiments, weekly traces, and small utilities from ani potts.",
+  hero_title: "making",
+  hero_summary:
+    "work i built or helped ship. product surfaces, data systems, quant tools, and the older pieces that still explain how i think.",
+};
+
+export const DEFAULT_PROJECTS_INDEX_CONTENT: ListingPageContent = {
+  title: "projects",
+  description: "every project with a page. flat list.",
+  hero_title: "archive",
+  hero_summary: "every project with a page. the curated story is at",
+  hero_link_label: "/making",
+  hero_link_href: "/making",
+};
+
+export const DEFAULT_NEWSLETTER_ARCHIVE_CONTENT: ListingPageContent = {
+  title: "newsletter archive",
+  description: "published issues from ani's first-party newsletter.",
+  hero_title: "newsletter archive",
+  hero_summary:
+    "published issues will appear here after the first-party send path is verified.",
+  section_label: "archive",
+};
+
+export const DEFAULT_ORCHESTRATING_CONTENT: OrchestratingPageContent = {
+  title: "orchestrating",
+  description:
+    "personal systems, labs, experiments, and local automation around ani's work.",
+  section_label: "orchestrating",
+  hero_title: "weird operating room",
+  hero_summary:
+    "where the private machinery goes when it is useful to see: local logs, small checks, cron traces, admin surfaces, and experiments that do not need a whole product around them.",
+  panel_label: "machine noise",
+  panel_copy: "tool calls captured from local session logs.",
+};
+
+export const DEFAULT_CMS_PROJECTS: CmsProjectContent[] = [
+  {
+    slug: "quantercise",
+    title: "quantercise",
+    status: "live",
+    year: "2024-",
+    range: "Ongoing",
+    tags: ["next.js", "typescript", "postgres", "drizzle", "stripe", "python"],
+    summary:
+      "quant prep with postgres, drizzle, stripe, and sandboxed python grading.",
+    body: "Built a quant interview prep app with 400+ problems, a Python editor, KaTeX math rendering, instant grading, and user progress. Next.js, TypeScript, Postgres, Drizzle, Stripe, and sandboxed Python grading.",
+    links: [{ label: "live site", url: "https://quantercise.com" }],
+    featured: true,
+    order: 100,
+    visible: true,
+  },
+  {
+    slug: "pgi-research-platform",
+    title: "pgi research portal",
+    status: "live",
+    year: "2025",
+    range: "2025-",
+    tags: ["next.js", "typescript", "supabase", "tailwindcss", "research"],
+    summary: "next.js and supabase portal for an nyu quant fund.",
+    body: "Built a Next.js and TypeScript research portal backed by Supabase for Paragon Global Investments, NYU's quant fund. Organized internal research so members could read and share fund material from mobile.",
+    links: [{ label: "live site", url: "https://paragoninvestments.org" }],
+    featured: false,
+    order: 95,
+    visible: true,
+  },
+  {
+    slug: "claude-code-tips",
+    title: "claude code tips",
+    status: "live",
+    year: "2025-",
+    range: "Ongoing",
+    tags: ["python", "claude code", "hooks", "agents", "mcp", "plugins"],
+    summary: "public notes from running agent workflows in real repos.",
+    body: "Reference repo built from hundreds of Claude Code sessions. Includes hooks, custom agents, MCP servers, plugin notes, and automation patterns that make agent work easier to debug.",
+    links: [
+      { label: "source", url: "https://github.com/anipotts/claude-code-tips" },
+    ],
+    featured: false,
+    order: 92,
+    visible: true,
+  },
+  {
+    slug: "imessage-mcp",
+    title: "imessage mcp",
+    status: "live",
+    year: "2025-",
+    range: "Ongoing",
+    tags: ["typescript", "mcp", "sqlite", "macos", "privacy"],
+    summary: "local-first mcp over imessage search and stats.",
+    body: "Built a macOS MCP server that lets local agents query iMessage history without writing to chat.db. Supports conversation search, contact stats, streaks, and local-only analytics. Published on npm as a small local-first tool.",
+    links: [
+      { label: "source", url: "https://github.com/anipotts/imessage-mcp" },
+      { label: "live site", url: "https://npmjs.com/package/imessage-mcp" },
+    ],
+    featured: true,
+    order: 91,
+    visible: true,
+  },
+  {
+    slug: "quantercise-extension",
+    title: "mental math extension",
+    status: "live",
+    year: "2026",
+    range: "Winter 2026",
+    tags: ["chrome extension", "javascript", "manifest v3"],
+    summary: "browser drills for fast mental math reps.",
+    body: "Built a zero-dependency Chrome extension for keyboard-driven mental math practice. Includes sound feedback, progress tracking, Manifest V3 packaging, and no external services.",
+    links: [
+      {
+        label: "source",
+        url: "https://github.com/anipotts/quantercise-mental-math-extension",
+      },
+    ],
+    featured: false,
+    order: 90,
+    visible: true,
+  },
+  {
+    slug: "chainedchat",
+    title: "chainedchat",
+    status: "archived",
+    year: "2025",
+    range: "Summer 2025",
+    tags: ["next.js", "typescript", "convex", "langgraph", "stripe"],
+    summary: "sunsetted multi-model chat with shared context and routing.",
+    body: "Archived Convex and Clerk app for running a conversation across multiple LLMs without losing context. It had shared-context caching, model routing, prompt-chain UI, and Stripe billing. Useful as a snapshot, not an active product.",
+    links: [
+      { label: "source", url: "https://github.com/anipotts/chained-chat" },
+    ],
+    featured: false,
+    order: 89,
+    visible: true,
+  },
+  {
+    slug: "saeshify",
+    title: "saeshify",
+    status: "live",
+    year: "2025-",
+    range: "Ongoing",
+    tags: ["typescript", "webgl", "audio", "visualization"],
+    summary: "real-time rhyme visualizer for lyrics and audio.",
+    body: "Built a WebGL and Web Audio pipeline that follows lyrics as they play and makes rhyme patterns visible in real time. Focused on timing, rendering, and keeping the interface fast enough to feel live.",
+    links: [
+      { label: "source", url: "https://github.com/anipotts/saeshify" },
+      { label: "live site", url: "https://saeshify.vercel.app" },
+    ],
+    featured: false,
+    order: 85,
+    visible: true,
+  },
+  {
+    slug: "nyu-purity-test",
+    title: "nyu purity test",
+    status: "live",
+    year: "2024",
+    range: "Fall 2024",
+    tags: ["typescript", "react", "next.js", "tailwindcss", "analytics"],
+    summary: "campus quiz with 3,000+ student completions.",
+    body: "Built and launched a TypeScript campus quiz in one night. It reached 1,000+ completions in under 17 hours and 200k+ visits through NYU group chats and social sharing.",
+    links: [{ label: "live site", url: "https://nyupuritytest.com" }],
+    featured: false,
+    order: 84,
+    visible: true,
+  },
+  {
+    slug: "habittracker-obh",
+    title: "artist scouting dashboard",
+    status: "live",
+    year: "2024",
+    range: "Summer 2024",
+    tags: ["python", "streamlit", "sql", "apis", "data pipelines"],
+    summary: "growth tracking for an atlantic records venture.",
+    body: "Built a scouting dashboard that pulled Chartmetric, YouTube, TikTok, and Instagram signals into one workflow. Added geo-based discovery for finding emerging artists in campaign markets.",
+    links: [],
+    featured: false,
+    order: 41,
+    visible: false,
+  },
+  {
+    slug: "options-pricing-sensitivity",
+    title: "options pricing + sensitivity analysis",
+    status: "live",
+    year: "2023",
+    range: "Spring 2023",
+    tags: ["python", "numpy", "pandas", "quantitative finance"],
+    summary: "black-scholes and binomial pricing with volatility sweeps.",
+    body: "Wrote a Python tool for pricing European options with Black-Scholes and binomial models, then swept volatility to show how model behavior changes.",
+    links: [
+      {
+        label: "source",
+        url: "https://github.com/anirudhp15/Options-Pricing-and-Sensitivity-Analysis-Tool",
+      },
+    ],
+    featured: false,
+    order: 40,
+    visible: true,
+  },
+];
+
+export const DEFAULT_CMS_WRITING: CmsWritingContent[] = [
+  {
+    slug: "saturdays-are-for-claude-code",
+    title: "saturdays are for claude code",
+    date: "2026-04-13",
+    tags: ["claude-code", "press", "workflow", "building"],
+    preview:
+      "business insider interviewed me about ai usage limits. the useful part was less the quote and more the workflow it forced.",
+    body: "A reporter from Business Insider reached out a couple weeks ago. He was writing about how usage limits on AI tools are changing the way people work. Somebody pointed him to me because I've been pretty vocal about how I use Claude Code. The article went live today.\n\nHe nailed the broad strokes. I do plan my work around session limits. I do save the hardest tasks for when I'm far from the cap. And yes, Saturdays are for Claude Code. That quote is real. My friends think I'm joking when I say that. I'm not.",
+    sourceLinks: [
+      {
+        label: "source",
+        url: "https://www.businessinsider.com/ai-usage-limits-causing-some-to-restructure-their-workday-2026-4",
+      },
+    ],
+    visible: true,
+    order: 50,
+  },
+  {
+    slug: "i-built-a-monitor-for-my-claude-code-sessions",
+    title: "i built a monitor for my claude code sessions",
+    date: "2026-04-07",
+    tags: ["claude-code", "claudemon", "observability", "building"],
+    preview:
+      "claude code has no dashboard for parallel sessions, so i'm building claudemon.",
+    body: "At any given time I've got like 3-5 Claude Code sessions running across different projects. One on a side project, one doing a refactor I kicked off before dinner, one on something else I already forgot about. And I have no idea what any of them are doing.\n\nClaude Code doesn't have a dashboard. There's no \"what are all your agents doing\" view. You get one terminal per session. If you're not staring at it, you're blind.\n\nSo I started building something. I'm calling it Claudemon.",
+    sourceLinks: [],
+    visible: true,
+    order: 49,
+  },
+  {
+    slug: "stop-ending-your-day-with-fix-the-bug",
+    title: 'stop ending your day with "fix the bug"',
+    date: "2026-04-07",
+    tags: ["claude-code", "productivity", "ai-tools"],
+    preview:
+      "vague todos waste context. specific prompts let claude code start from the right file.",
+    body: 'I used to end my day with todos like "fix auth" and "clean up API" and then wake up the next morning having no idea what I actually meant.\n\nFix auth how? Which auth? The login flow? The token refresh? The middleware? I\'d spend the first 20 minutes of my next session just rebuilding the context I had the night before.\n\nThis is 10x worse with Claude Code. When you hand a vague todo to an AI coding agent, it doesn\'t just lose context. It actively goes searching for context.',
+    sourceLinks: [],
+    visible: true,
+    order: 48,
+  },
+  {
+    slug: "jpegmafia-is-our-kanye-west",
+    title: "jpegmafia is our kanye west",
+    date: "2026-02-12",
+    tags: ["music", "product", "execution"],
+    preview:
+      "a short note on taste, consensus, feedback, and shipping with strong defaults.",
+    body: "Teams often over-index on being understood early\n\nThe highest leverage work often looks wrong in public before it looks obvious in hindsight\n\nJPEGMAFIA reminds me of that builder pattern\n\n- hard constraints\n- aggressive iteration\n- distinct taste\n- zero permission waiting\n\nConsensus can improve distribution, but it rarely creates a category-defining product",
+    sourceLinks: [],
+    visible: true,
+    order: 47,
+  },
+  {
+    slug: "search-will-be-dead-by-2030",
+    title: "search will be dead by 2030",
+    date: "2026-01-31",
+    tags: ["ai", "search", "product"],
+    preview:
+      "search does not disappear. the main interface collapses into agents, memory, retrieval, and synthesis.",
+    body: "Classic search assumes users can translate intent into keywords\n\nThe next interface assumes the system already understands intent from context and history\n\nWhat changes first:\n\n- query boxes become fallback UX\n- ranking systems become orchestration systems\n- links become evidence, not destination\n\nSearch is not dead as infrastructure\n\nSearch is dead as a primary interaction model",
+    sourceLinks: [],
+    visible: true,
+    order: 46,
+  },
+];
