+++ packages/lib/package.json
@@ -95,6 +95,7 @@
     "test": "vitest run"
   },
   "dependencies": {
+    "@anipotts/content": "workspace:*",
     "@anipotts/types": "workspace:*",
     "clsx": "^2.1.1",
     "drizzle-orm": "^0.45.2",
